Improve tracing spans and events
* Spans now generated for every packet and timeout event
* Ephemeral "startup" span runs before main loop kicks off
* `self.state` and `self.term` logged on parent spans;
  removed from individual event fields

n.b. There is a bug in tracing_subscriber that causes the
term field to be logged twice when an election timeout
occurs in Candidate state. The `tracing` crate correctly
updates the original `term` field on the span to the new
value, but `tracing_subscriber` only appends new values
to the log line for... reasons. See:
https://github.com/tokio-rs/tracing/issues/2334