Move client response await into ServerHandle
Server shouldn't await responses, because it might
change state while a client is waiting, and wouldn't have exclusive
ownership over itself to be able to convert to the next state.

New design passes the Receiver back to the ServerHandle, where
it can be awaited until either 1) success, 2) timeout, or 3) Server
drops the Sender half because it is changing state.

Also, return an error if a client request is given to a Follower
or a Candidate instead of a Leader. The Follower error is to redirect
to the leader. The Candidate error is to say that the cluster
is in an un-led state and to retry later. These errors return
the client value so that it can be retried, which required making
ServerError into ServerError<V: JournalValue>.