Change commit index to an Option<u64>
Places in the code were not sure whether to act as if commit_index
was zero, the first entry in the journal, or zero, meaning no committed
entries. Update the code base to pass around Option<u64> for commit
index, to be certain which state we are in. Not storing the option
directly in `Journal`, because it would require a Mutex<Option<u64>>.
Instead implementing it as a pair of AtomicUsize and AtomicBool.