Journal compaction
Journal is now snapshotted on a regular basis,
and prior entries and snapshots deleted, to maintain consistent
length. When followers are restarted or fall out of sync,
a single snapshot, nearest to the follower's match_index,
is sent to catch it up.

Implementing required a switch from std::sync::RwLock to
tokio::sync::RwLock in Journal and a change from sync
to async functions in the Journal trait.

Also made many changes to eliminate lock contention by
making helper methods accept a lock guard instead of trying
to obtain their own locks. Deadlocks had been possible if
a helper method tried to obtain a read lock when called from
a method that already held a read lock, if another thread
got into the lock queue for a write lock, because the helper's
read lock would not be granted before the other thread's write
lock--which could not be granted until the original caller's
read lock was dropped.